+++ src/product_pages/index.html
@@ -60,7 +60,7 @@
 
         <h2 class="divider" id="productName"></h2>
 
-        <img id="productImage" class="divider" src="" alt="" />
+        <img id="productImg" class="divider" src="" alt="" />
 
         <p class="product-card__price" id="productFinalPrice"></p>
